Use tmux-sensible defaults for tmux shell
In 53a09c5d4d6a172cc6b0bcd9c268513afba0c5d8 I had changed the
default-command to "$SHELL" to enforce a non-login shell, because I had
made config.fish only source the shared config in login shells. This is
no longer the case: config.fish always sources .profile (using fenv,
faster that bass which I had used at the time). The other reason for
going back is that tmux sensible will use reattach-to-user-namespace if
available. This is not needed for the clipboard anymore, but it is
needed for the "open" command. Now I get this:

    LSOpenURLsWithRole() failed with error -600 for the file ...

This also changes the Pageup/Pagedown bindings to send them unchanged to
vim, since they work fine (instead of C-b/C-f).